third_party/WebKit/LayoutTests/media/audio-controls-do-not-fade-out.html

<script src="media-controls.js"></script>
<body>
<p>
    This tests that audio controls do not fade out when the audio is playing.
</p>
<p id="result">
    FAIL: Test did not run.
</p>
<audio id="audio" controls autoplay src="content/test.oga"></audio><br>
<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText();
    }

    document.getElementById("audio").addEventListener("timeupdate", function(e)
    {
        if (e.target.currentTime < 4) {
            return;
        }
        var controls = mediaControlsElement(internals.shadowRoot(e.target).firstChild, "-webkit-media-controls-panel");
        var opacity = getComputedStyle(controls).opacity;
        document.getElementById("result").innerText = opacity < 1 ? "FAIL" : "PASS";
        testRunner.notifyDone();
    });
</script>
</body>
